+++ packages/mobile-dev-tools/src/index.ts
@@ -1,3 +1,3 @@
-export const VERSION = "0.10.0";
+export const VERSION = "0.11.0";
 
 export const PACKAGE_NAME = "@tmhs/mobile-dev-tools";